requirements.txt
asgiref==3.8.1
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
mysqlclient==2.2.4
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
drf-spectacular
pytest
pytest-django